requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
openai==0.28.0
langchain>=0.0.310
langchain-core>=0.1.7,<0.2.0
langgraph>=0.0.19 